refactor(vscode): isolate node version helpers
